Refactor blank project modules for repo standards
Problem:
Repository lint rejected growth in four already oversized source and test files.

Root cause:
The first implementation placed new backend logic, tests, bridge methods, and project creation state directly in ratcheted legacy files.

Fix:
Extract the blank-project backend, tests, browser bridge, and project-creation state into focused modules while keeping the public behavior unchanged.

Verification:
- go run ./tools/repolint
- go test -race ./desktop -run "TestCreateBlankProject|TestSwitchWorkspaceRegistersDefaultTopicInProjectTree" -count=1
- pnpm exec tsx src/__tests__/blank-project-dialog.test.tsx
- pnpm exec tsx src/__tests__/project-tree-runtime.test.ts
- pnpm typecheck
- pnpm test:typecheck
- pnpm build
