+++ desktop/frontend/src/components/ProjectTree.tsx
@@ -2,7 +2,7 @@
 // It shows a tree of projects (each with expandable topics) plus a Global
 // section. Clicking a topic opens its tab; "+" next to a project creates a
 // new topic.
-import { lazy, Suspense, useCallback, useEffect, useMemo, useRef, useState } from "react";
+import { useCallback, useEffect, useMemo, useRef, useState } from "react";
 import { createPortal } from "react-dom";
 import type { CSSProperties, DragEvent as ReactDragEvent, KeyboardEvent as ReactKeyboardEvent, MouseEvent as ReactMouseEvent } from "react";
 import { Archive, ArrowDown, Pencil, Plus, Folder, FolderPlus, Search, BriefcaseBusiness, Copy, FolderOpen, XCircle, Check, ListCollapse, ListRestart, MessageSquare, Clock, Pin, MoreHorizontal, Minimize2, Maximize2, GitBranch } from "lucide-react";
@@ -22,8 +22,7 @@ import type { ShortcutPlatform } from "../lib/keyboardShortcuts";
 import { ContextMenu, contextMenuPointFromEvent, type ContextMenuItem, type ContextMenuPoint } from "./ContextMenu";
 import { Tooltip } from "./Tooltip";
 import { WorktreeBadge } from "./WorktreeBadge";
-
-const BlankProjectFlow = lazy(() => import("./BlankProjectFlow").then((module) => ({ default: module.BlankProjectFlow })));
+import { useProjectCreation } from "./useProjectCreation";
 
 interface ProjectTreeProps {
   activeScope?: string;
@@ -434,8 +433,6 @@ export function ProjectTree({
   const [menuPoint, setMenuPoint] = useState<ContextMenuPoint | null>(null);
   const [editingProject, setEditingProject] = useState<{ key: string; root: string } | null>(null);
   const [projectDraft, setProjectDraft] = useState("");
-  const [addingProject, setAddingProject] = useState(false);
-  const [blankProjectFlowOpen, setBlankProjectFlowOpen] = useState(false);
   const [isolatingProject, setIsolatingProject] = useState<string | null>(null);
   const [worktreeAvailability, setWorktreeAvailability] = useState<Record<string, { available: boolean; reason?: string }>>({});
   const [confirmAction, setConfirmAction] = useState<{ topicId: string; action: "trash" } | null>(null);
@@ -568,6 +565,11 @@ export function ProjectTree({
       /* bridge unavailable */
     }
   }, []);
+  const { addingProject, handleAddProject, openBlankProjectFlow, blankProjectFlow } = useProjectCreation({
+    onAddProject,
+    onRefresh: refresh,
+    showToast,
+  });
 
   useEffect(() => {
     treeRef.current = tree;
@@ -793,19 +795,6 @@ export function ProjectTree({
       return changed ? next : prev;
     });
   }, [collapseSnapshot, expanded, folderKeys, hasExpandedFolders, manuallyCollapsed, searchActive, updateManuallyCollapsed]);
-
-  const handleAddProject = async () => {
-    if (addingProject) return;
-    setAddingProject(true);
-    try {
-      await onAddProject();
-      await refresh();
-    } catch (err) {
-      showToast(err instanceof Error ? err.message : String(err), "error");
-    } finally {
-      setAddingProject(false);
-    }
-  };
 
   const openWorkbenchHeaderMenu = (
     event: ReactMouseEvent<HTMLElement> | ReactKeyboardEvent<HTMLElement>,
@@ -1899,10 +1888,7 @@ export function ProjectTree({
       icon: <FolderPlus size={13} />,
       label: t("projectTree.createBlankProject"),
       disabled: addingProject,
-      onSelect: () => {
-        closeMenu();
-        setBlankProjectFlowOpen(true);
-      },
+      onSelect: () => { closeMenu(); openBlankProjectFlow(); },
     },
     {
       key: "existing-folder",
@@ -2278,15 +2264,7 @@ export function ProjectTree({
         </div>,
         document.body,
       )}
-      {blankProjectFlowOpen && (
-        <Suspense fallback={null}>
-          <BlankProjectFlow
-            onOpenProject={(path) => onAddProject(path)}
-            onRefresh={() => refresh()}
-            onClose={() => setBlankProjectFlowOpen(false)}
-          />
-        </Suspense>
-      )}
+      {blankProjectFlow}
     </div>
   );
 }
